+++ packages/dts-gen/jest.config.js
@@ -1,4 +1,7 @@
-module.exports = {
-    rootDir: "src",
-    testEnvironment: "node",
+/** @type {import('@jest/types').Config.InitialOptions} */
+const config = {
+  roots: ["<rootDir>/src"],
+  testEnvironment: "node",
+  testTimeout: 30000,
 };
+module.exports = config;