+++ packages/dts-gen/src/index.ts
@@ -2,7 +2,7 @@ import { FormsClientImpl } from "./kintone/clients/forms-client-impl";
 import { DemoClient } from "./kintone/clients/demo-client";
 import { FieldTypeConverter } from "./converters/fileldtype-converter";
 import { TypeDefinitionTemplate } from "./templates/template";
-import { objectValues } from "./utils//objectvalues";
+import { objectValues } from "./utils/objectvalues";
 import { parse } from "./cli-parser";
 
 process.on("uncaughtException", (e) => {
@@ -21,22 +21,17 @@ const fetchFormPropertiesInput = {
   preview: args.preview,
 };
 
-client
-  .fetchFormProperties(fetchFormPropertiesInput)
-  .then((properties) =>
-    FieldTypeConverter.convertFieldTypesToFieldTypeGroups(
-      objectValues(properties)
-    )
-  )
-  .then((fieldTypeGroups) => {
-    const typeName = args.typeName;
-    const namespace = args.namespace;
-    const input = {
-      typeName,
-      namespace,
-      fieldTypeGroups,
-    };
-    TypeDefinitionTemplate.renderAsFile(args.output, input);
-  })
-  // eslint-disable-next-line no-console
-  .catch((err) => console.error(err));
+const handler = async () => {
+  const properties = await client.fetchFormProperties(fetchFormPropertiesInput);
+  const fieldTypeGroups = FieldTypeConverter.convertFieldTypesToFieldTypeGroups(
+    objectValues(properties)
+  );
+  const input = {
+    typeName: args.typeName,
+    namespace: args.namespace,
+    fieldTypeGroups,
+  };
+  await TypeDefinitionTemplate.renderAsFile(args.output, input);
+};
+
+handler().catch((err) => console.error(err));