+++ packages/plugin-packer/site/jest.config.js
@@ -1,4 +1,6 @@
-module.exports = {
+/** @type {import('@jest/types').Config.InitialOptions} */
+const config = {
   testRegex: "/test/.*-test\\.js$",
   testEnvironment: "jsdom",
 };
+module.exports = config;