+++ packages/profile-loader/jest.config.js
@@ -1,4 +1,6 @@
-module.exports = {
-  rootDir: "src",
+/** @type {import('@jest/types').Config.InitialOptions} */
+const config = {
+  roots: ["<rootDir>/src"],
   testEnvironment: "node",
 };
+module.exports = config;