+++ packages/plugin-packer/CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.0.3](https://github.com/kintone/js-sdk/compare/@kintone/plugin-packer@6.0.2...@kintone/plugin-packer@6.0.3) (2022-06-08)
+
+**Note:** Version bump only for package @kintone/plugin-packer
+
+
+
+
+
+## [6.0.2](https://github.com/kintone/js-sdk/compare/@kintone/plugin-packer@6.0.1...@kintone/plugin-packer@6.0.2) (2022-06-01)
+
+**Note:** Version bump only for package @kintone/plugin-packer
+
+
+
+
+
+## [6.0.1](https://github.com/kintone/js-sdk/compare/@kintone/plugin-packer@6.0.0...@kintone/plugin-packer@6.0.1) (2022-05-20)
+
+**Note:** Version bump only for package @kintone/plugin-packer
+
+
+
+
+
+# [6.0.0](https://github.com/kintone/js-sdk/compare/@kintone/plugin-packer@5.0.39...@kintone/plugin-packer@6.0.0) (2022-05-13)
+
+
+* chore!: drop Node v12 support because of the EOL (BREAKING CHANGE) (#1493) ([0d9dae1](https://github.com/kintone/js-sdk/commit/0d9dae10582fc40d89a1af8db4a2efc1d776a456)), closes [#1493](https://github.com/kintone/js-sdk/issues/1493)
+
+
+### BREAKING CHANGES
+
+* drop Node v12 support because of the EOL.
+
+* ci: update Node version 14 -> 16
+
+* ci: remove Node 18.x from test workflow
+
+
+
+
+
 ## [5.0.39](https://github.com/kintone/js-sdk/compare/@kintone/plugin-packer@5.0.38...@kintone/plugin-packer@5.0.39) (2022-04-28)
 
 **Note:** Version bump only for package @kintone/plugin-packer